+++ packages/tssdk/src/index.ts
@@ -10,6 +10,8 @@ export {
   EvmContract,
 } from './libs';
 
+export { msgPackEncoder } from './utils/msgpack';
+
 export { ChainAction } from './helpers/network.enum';
 export { NetworkEnum } from './config/network.enum';
 export * from './typings';